debounce the squiggles processing
